requirements.txt
fastapi
langchain
langchain-community
langchain-core
langchain-ollama
langchain-openai
langchain-aws
litserve
numpy
openai
pillow
pydantic
regex
requests
wandb
weave
streamlit
fpdf
langgraph